test: consolidate prefer-array-flat coverage
