+++ internal/plugins/unicorn/rules/prefer_array_flat/prefer_array_flat_extras_test.go
@@ -399,6 +399,96 @@ func TestPreferArrayFlatExtras(t *testing.T) {
 		map[string]interface{}{},
 	)
 
+	// ---- Dimension 3: comment ownership and fix suppression ----
+	suite.addNoFix(
+		`array.flatMap(value /* keep */ => value)`,
+		`array.flatMap(value /* keep */ => value)`,
+		`Array#flatMap()`,
+		nil,
+	)
+	suite.addFixed(
+		`foo./* keep */bar.flatMap(value => value)`,
+		`foo./* keep */bar.flatMap(value => value)`,
+		`foo./* keep */bar.flat()`,
+		`Array#flatMap()`,
+		nil,
+	)
+
+	// ---- Dimension 3: member-object parentheses ----
+	expressionCases := []struct {
+		code        string
+		replacement string
+	}{
+		{code: `_.flatten({value: 1})`, replacement: `({value: 1}).flat()`},
+		{code: `_.flatten(class {})`, replacement: `(class {}).flat()`},
+		{code: `_.flatten(value => value)`, replacement: `(value => value).flat()`},
+		{code: `_.flatten(flag ? left : right)`, replacement: `(flag ? left : right).flat()`},
+		{code: `_.flatten(value = input)`, replacement: `(value = input).flat()`},
+		{code: `_.flatten(++value)`, replacement: `(++value).flat()`},
+		{code: `_.flatten("value")`, replacement: `"value".flat()`},
+		{code: `_.flatten(/value/)`, replacement: `/value/.flat()`},
+		{code: `_.flatten(1n)`, replacement: `1n.flat()`},
+		{code: `_.flatten(true)`, replacement: `true.flat()`},
+		{code: `_.flatten(null)`, replacement: `null.flat()`},
+		{code: "_.flatten(`value`)", replacement: "`value`.flat()"},
+	}
+	for _, testCase := range expressionCases {
+		suite.addFixed(
+			testCase.code,
+			testCase.code,
+			testCase.replacement,
+			`_.flatten()`,
+			nil,
+		)
+	}
+	suite.addFixed(
+		`function* values() { return _.flatten(yield input); }`,
+		`_.flatten(yield input)`,
+		`(yield input).flat()`,
+		`_.flatten()`,
+		nil,
+	)
+
+	// ---- Dimension 3: ASI-sensitive embedded statement bodies ----
+	embeddedCases := []string{
+		`while (condition)
+	Array.prototype.concat.call([], value)`,
+		`for (;;)
+	Array.prototype.concat.call([], value)`,
+		`for (const item of items)
+	Array.prototype.concat.call([], value)`,
+		`do
+	Array.prototype.concat.call([], value)
+while (condition)`,
+		`with (object)
+	Array.prototype.concat.call([], value)`,
+	}
+	for _, code := range embeddedCases {
+		suite.addFixed(
+			code,
+			`Array.prototype.concat.call([], value)`,
+			`[value].flat()`,
+			`Array.prototype.concat()`,
+			nil,
+		)
+	}
+	suite.addFixed(
+		`before()
+Array.prototype.concat.call([], value)`,
+		`Array.prototype.concat.call([], value)`,
+		`;[value].flat()`,
+		`Array.prototype.concat()`,
+		nil,
+	)
+	suite.addFixed(
+		`value = {}
+Array.prototype.concat.call([], item)`,
+		`Array.prototype.concat.call([], item)`,
+		`;[item].flat()`,
+		`Array.prototype.concat()`,
+		nil,
+	)
+
 	// N/A: string/numeric/private object or class property keys are not rule
 	// inputs; only call callee member accesses are inspected.
 	// N/A: class declaration/expression and function declaration/expression
